Change float to double to avoid buffer overflow, since a double is read into this var.
git-svn-id: http://svn.jcvi.org/ANNOTATION/proteogenomics/trunk@30082 061605af-e407-0410-a945-9c8a22772c06

+++ LDA.c
@@ -129,7 +129,7 @@ LDAModel* LoadLDAModel(char* LDAModelFileName)
 {
     FILE* File;
     LDAModel* Model;
-    float Value;
+    double Value;
     int BytesRead;
     //
     File = fopen(LDAModelFileName, "rb");
